Renamed base folder to sprites https://github.com/achobanov/Web-Game/issues/11
