+++ Source/Client/source/objects/sprites/entity.ts
